Refactor into multiple artifacts
