--- visualization/projection/README.md
@@ -1,2 +1,0 @@
-# visualization
-Simple visualization serializer - convert one or more containers to a JSON file for use in the viewer.
