--- visualization/projection/src/main/java/com/github/skjolber/packing/visualization/BoxVisualization.java
@@ -1,14 +1,0 @@
-package com.github.skjolber.packing.visualization;
-
-public class BoxVisualization extends StackableVisualization {
-
-	private String type = "box";
-
-	public String getType() {
-		return type;
-	}
-	
-	public void setType(String type) {
-		this.type = type;
-	}
-}
