--- visualization/projection/src/main/java/com/github/skjolber/packing/visualization/StackableVisualization.java
@@ -1,28 +1,0 @@
-package com.github.skjolber.packing.visualization;
-
-public class StackableVisualization extends AbstractVisualization {
-
-	private int dx;
-	private int dy;
-	private int dz;
-
-	public int getDx() {
-		return dx;
-	}
-	public void setDx(int dx) {
-		this.dx = dx;
-	}
-	public int getDy() {
-		return dy;
-	}
-	public void setDy(int dy) {
-		this.dy = dy;
-	}
-	public int getDz() {
-		return dz;
-	}
-	public void setDz(int dz) {
-		this.dz = dz;
-	}
-
-}
