--- visualization/projection/src/main/java/com/github/skjolber/packing/visualization/VisualizationPlugin.java
@@ -1,7 +1,0 @@
-package com.github.skjolber.packing.visualization;
-
-public interface VisualizationPlugin {
-
-	String getId();
-	
-}
