--- visualization/projection/src/main/java9/module-info.java
@@ -1,3 +1,0 @@
-module com.github.skjolber.packing.visualization {
-	exports com.github.skjolber.packing.visualization;
-}
