+++ visualizer/api/README.md
@@ -1,0 +1,5 @@
+# visualizer api
+API artifact for JSON formats:
+
+ * packaging results (i.e. list of containers)
+ * algorithm results (i.e. steps and evaluations)
