+++ visualizer/api/src/main/java/com/github/skjolber/packing/projection/BoxVisualization.java
@@ -1,0 +1,14 @@
+package com.github.skjolber.packing.projection;
+
+public class BoxVisualization extends StackableVisualization {
+
+	private String type = "box";
+
+	public String getType() {
+		return type;
+	}
+	
+	public void setType(String type) {
+		this.type = type;
+	}
+}
