+++ visualizer/api/src/main/java/com/github/skjolber/packing/projection/StackPlacementVisualization.java
@@ -1,0 +1,37 @@
+package com.github.skjolber.packing.projection;
+
+public class StackPlacementVisualization extends AbstractVisualization {
+
+	private int x;
+	private int y;
+	private int z;
+	
+	private StackableVisualization stackable;
+
+	public int getX() {
+		return x;
+	}
+	public void setX(int x) {
+		this.x = x;
+	}
+	public int getY() {
+		return y;
+	}
+	public void setY(int y) {
+		this.y = y;
+	}
+	public int getZ() {
+		return z;
+	}
+	public void setZ(int z) {
+		this.z = z;
+	}
+
+	public void setStackable(StackableVisualization stackable) {
+		this.stackable = stackable;
+	}
+
+	public StackableVisualization getStackable() {
+		return stackable;
+	}
+}
