+++ visualizer/api/src/main/java/com/github/skjolber/packing/projection/StackableVisualization.java
@@ -1,0 +1,28 @@
+package com.github.skjolber.packing.projection;
+
+public class StackableVisualization extends AbstractVisualization {
+
+	private int dx;
+	private int dy;
+	private int dz;
+
+	public int getDx() {
+		return dx;
+	}
+	public void setDx(int dx) {
+		this.dx = dx;
+	}
+	public int getDy() {
+		return dy;
+	}
+	public void setDy(int dy) {
+		this.dy = dy;
+	}
+	public int getDz() {
+		return dz;
+	}
+	public void setDz(int dz) {
+		this.dz = dz;
+	}
+
+}
