+++ visualizer/api/src/main/java/com/github/skjolber/packing/projection/VisualizationPlugin.java
@@ -1,0 +1,7 @@
+package com.github.skjolber.packing.projection;
+
+public interface VisualizationPlugin {
+
+	String getId();
+	
+}
