+++ visualizer/packaging/README.md
@@ -1,0 +1,2 @@
+# packaging visualizer
+Simple visualization of packaging results. This module convert one or more containers to a JSON file for use in the viewer.
