+++ visualizer/packaging/src/main/java9/module-info.java
@@ -1,0 +1,3 @@
+module com.github.skjolber.packing.visualization {
+	exports com.github.skjolber.packing.visualizer.packaging;
+}
