+++ visualizer/viewer/public/assets/containers.json
